Creators cannot depend directly on CBD dest nodes
If a destroy node is CreateBeforeDestroy, we need to remove any direct
dependencies on it from resource creation nodes to prevent cycles.

Since a create node cannot both depend on its destroy node AND be
CreateBeforeDestroy, the same goes for its dependencies. While we do
connect resources with dependency destroy nodes so that updates are
ordered correctly, this ordering does not make sense in the
CreateBeforeDestroy case.